+++ idea/frontend/src/components/blocks/ProgramForm/ProgramForm.tsx
@@ -9,7 +9,7 @@ import { INITIAL_VALUES } from './const';
 import { getValidationSchema } from './Schema';
 import { FormValues, PropsToRenderButtons, SetFieldValue, Helpers } from './types';
 
-import { Payload } from 'hooks/useProgramUplaod/types';
+import { Payload } from 'hooks/useProgramActions/types';
 import { getSubmitPayload, getPayloadFormValues } from 'components/common/Form/FormPayload/helpers';
 import { Fieldset } from 'components/common/Fieldset';
 import { FormInput, FormPayload, FormPayloadType, FormNumberFormat, formStyles } from 'components/common/Form';
