+++ idea/frontend/src/hooks/useProgramActions/helpers.ts
