+++ idea/frontend/src/hooks/useProgramActions/index.ts
@@ -1,2 +1,2 @@
 // eslint-disable-next-line import/no-cycle
-export * from './useProgramUpload';
+export * from './useProgramActions';
